requirements.txt
asn1crypto==0.24.0
backports.functools-lru-cache==1.6.1
beautifulsoup4==4.9.1
cachetools==3.1.1
certifi==2020.4.5.1
chardet==3.0.4
configparser==3.5.0b2
cryptography==2.6.1
dbus-python==1.2.12
django-bootstrap==0.2.4
django-bootstrap4==1.1.1
entrypoints==0.3
enum34==1.1.6
futures==3.3.0
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
gyp==0.1
httplib2==0.17.2
idna==2.9
ipaddress==1.0.17
keyring==18.0.1
keyrings.alt==3.1.1
protobuf==3.11.3
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
pycurl==7.43.0.2
PyGObject==3.34.0
pyliblzma==0.5.3
pysqlite==1.0.1
pytz==2019.3
pyxdg==0.25
redis==2.10.6
requests==2.23.0
rpm==4.14.2.1
rsa==4.0
SecretStorage==2.3.1
selenium==3.141.0
six==1.12.0
soupsieve==1.9.6
uritemplate==3.0.1
urlgrabber==3.10.2
urllib3==1.25.8
yum-metadata-parser==1.1.4
